refactor(api): move get api/admin/users/{id} in deprecated
