+++ api/src/deprecated/application/routes.js
@@ -1,9 +1,15 @@
 import { certificationPointOfContactRoute } from './certification-point-of-contact.route.js';
 import { prescriberInformationsRoute } from './prescriber-informations.route.js';
+import { userAdminRoutes } from './user-admin.route.js';
 import { usersMeRoute } from './users-me.route.js';
 
 const register = async function (server) {
-  server.route([...certificationPointOfContactRoute, ...prescriberInformationsRoute, ...usersMeRoute]);
+  server.route([
+    ...certificationPointOfContactRoute,
+    ...prescriberInformationsRoute,
+    ...usersMeRoute,
+    ...userAdminRoutes,
+  ]);
 };
 
 const name = 'deprecated/deprecated-api';
