+++ api/src/deprecated/dependencies.json
@@ -6,6 +6,7 @@
     "legal-documents",
     "organizational-entities",
     "prescription/campaign-participation",
+    "prescription/learner-management",
     "shared",
     "team"
   ],
