+++ api/src/deprecated/domain/models/UserDetailsForAdmin.js
