+++ api/src/deprecated/domain/usecases/get-user-details-for-admin.usecase.js
@@ -1,12 +1,13 @@
 /**
  * @param {Object} params
  * @param {string} params.userId
- * @param {UserRepository} userRepository
+ * @param {userAdminRepository} userAdminRepository
+ * @param {legalDocumentApiRepository} legalDocumentApiRepository
  * @throws UserNotFoundError
  * @returns {Promise<UserDetailsForAdmin>}
  */
-const getUserDetailsForAdmin = async function ({ userId, userRepository, legalDocumentApiRepository }) {
-  const userDetailsForAdmin = await userRepository.getUserDetailsForAdmin(userId);
+const getUserDetailsForAdmin = async function ({ userId, userAdminRepository, legalDocumentApiRepository }) {
+  const userDetailsForAdmin = await userAdminRepository.getUserDetailsForAdmin(userId);
 
   const pixAppTosStatus = await legalDocumentApiRepository.getPixAppTosStatus({ userId });
   const pixOrgaTosStatus = await legalDocumentApiRepository.getPixOrgaTosStatus({ userId });
