+++ api/src/deprecated/infrastructure/serializers/jsonapi/user-details-for-admin.serializer.js
