+++ api/src/identity-access-management/application/user/user.admin.controller.js
@@ -1,7 +1,6 @@
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { userAnonymizedDetailsForAdminSerializer } from '../../infrastructure/serializers/jsonapi/user-anonymized-details-for-admin.serializer.js';
-import { userDetailsForAdminSerializer } from '../../infrastructure/serializers/jsonapi/user-details-for-admin.serializer.js';
 import { userForAdminSerializer } from '../../infrastructure/serializers/jsonapi/user-for-admin.serializer.js';
 import { userLoginSerializer } from '../../infrastructure/serializers/jsonapi/user-login-serializer.js';
 import { userUpdateForAdminSerializer } from '../../infrastructure/serializers/jsonapi/user-update-for-admin.serializer.js';
@@ -48,20 +47,6 @@ const updateUserDetailsByAdmin = async function (request, h, dependencies = { us
 
   await usecases.updateUserDetailsByAdmin({ userId, userToUpdate, updatedByAdminId });
   return h.response().code(204);
-};
-
-/**
- *
- * @param request
- * @param h
- * @param dependencies
- * @param {UserDetailsForAdminSerializer} dependencies.userDetailsForAdminSerializer
- * @returns {Promise<*>}
- */
-const getUserDetails = async function (request, h, dependencies = { userDetailsForAdminSerializer }) {
-  const userId = request.params.id;
-  const userDetailsForAdmin = await usecases.getUserDetailsForAdmin({ userId });
-  return dependencies.userDetailsForAdminSerializer.serialize(userDetailsForAdmin);
 };
 
 /**
@@ -144,15 +129,12 @@ const reassignAuthenticationMethod = async function (request, h) {
  * @property {function} unblockUserAccount
  * @property {function} updateUserDetailsByAdmin
  */
-const userAdminController = {
+export const userAdminController = {
   addPixAuthenticationMethod,
   anonymizeUser,
   findPaginatedFilteredUsers,
-  getUserDetails,
   reassignAuthenticationMethod,
   removeAuthenticationMethod,
   unblockUserAccount,
   updateUserDetailsByAdmin,
 };
-
-export { userAdminController };
