+++ api/src/identity-access-management/infrastructure/repositories/user.repository.js
@@ -1,8 +1,5 @@
 import { InvalidOrAlreadyUsedEmailError } from '../../../identity-access-management/domain/errors.js';
-import * as organizationFeaturesApi from '../../../organizational-entities/application/api/organization-features-api.js';
 import { Organization } from '../../../organizational-entities/domain/models/Organization.js';
-import { OrganizationLearnerForAdmin } from '../../../prescription/learner-management/domain/read-models/OrganizationLearnerForAdmin.js';
-import * as organizationLearnerImportFormatRepository from '../../../prescription/learner-management/infrastructure/repositories/organization-learner-import-format-repository.js';
 import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 import {
   AlreadyExistingEntityError,
@@ -13,10 +10,7 @@ import { Membership } from '../../../shared/domain/models/Membership.js';
 import { fetchPage, isUniqConstraintViolated } from '../../../shared/infrastructure/utils/knex-utils.js';
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../domain/constants/identity-providers.js';
 import { QUERY_TYPES } from '../../domain/constants/user-query.js';
-import { LastUserApplicationConnection } from '../../domain/models/LastUserApplicationConnection.js';
 import { User } from '../../domain/models/User.js';
-import { UserDetailsForAdmin } from '../../domain/models/UserDetailsForAdmin.js';
-import { UserLogin } from '../../domain/models/UserLogin.js';
 
 const getByEmail = async function (email) {
   const knexConn = DomainTransaction.getConnection();
@@ -65,70 +59,6 @@ const getByIds = async function (userIds) {
   const dbUsers = await knexConn('users').whereIn('id', userIds);
 
   return dbUsers.map((dbUser) => new User(dbUser));
-};
-
-const getUserDetailsForAdmin = async function (userId) {
-  const knexConn = DomainTransaction.getConnection();
-  const userDTO = await knexConn('users')
-    .leftJoin('user-logins', 'user-logins.userId', 'users.id')
-    .leftJoin('users AS anonymisedBy', 'anonymisedBy.id', 'users.hasBeenAnonymisedBy')
-    .select([
-      'users.*',
-      'user-logins.id AS userLoginId',
-      'user-logins.failureCount',
-      'user-logins.temporaryBlockedUntil',
-      'user-logins.blockedAt',
-      'user-logins.lastLoggedAt AS lastLoggedAt',
-      'anonymisedBy.firstName AS anonymisedByFirstName',
-      'anonymisedBy.lastName AS anonymisedByLastName',
-    ])
-    .where({ 'users.id': userId })
-    .first();
-
-  if (!userDTO) {
-    throw new UserNotFoundError(`User not found for ID ${userId}`);
-  }
-
-  const lastUserApplicationConnectionsDTO = await knexConn('last-user-application-connections').where({ userId });
-
-  const authenticationMethodsDTO = await knexConn('authentication-methods')
-    .select([
-      'authentication-methods.id',
-      'authentication-methods.identityProvider',
-      'authentication-methods.authenticationComplement',
-      'authentication-methods.lastLoggedAt',
-    ])
-    .join('users', 'users.id', 'authentication-methods.userId')
-    .where({ userId });
-
-  const organizationLearnersDTO = await knexConn('view-active-organization-learners')
-    .select([
-      'view-active-organization-learners.*',
-      'organizations.name AS organizationName',
-      'organizations.isManagingStudents AS organizationIsManagingStudents',
-    ])
-    .join('organizations', 'organizations.id', 'view-active-organization-learners.organizationId')
-    .where({ userId })
-    .orderBy('id');
-
-  for (const learner of organizationLearnersDTO) {
-    const features = await organizationFeaturesApi.getAllFeaturesFromOrganization(learner.organizationId);
-    learner.hasImportFeature = features.hasLearnersImportFeature;
-    if (learner.hasImportFeature) {
-      const importFormat = await organizationLearnerImportFormatRepository.get(learner.organizationId);
-      learner.additionalColumns = importFormat.extraColumns;
-    }
-  }
-
-  const pixAdminRolesDTO = await knexConn('pix-admin-roles').where({ userId });
-
-  return _fromKnexDTOToUserDetailsForAdmin({
-    userDTO,
-    organizationLearnersDTO,
-    authenticationMethodsDTO,
-    pixAdminRolesDTO,
-    lastUserApplicationConnectionsDTO,
-  });
 };
 
 const findPaginatedFiltered = async function ({ filter, page, queryType = QUERY_TYPES.CONTAINS }) {
@@ -408,7 +338,6 @@ const updateLastDataProtectionPolicySeenAt = async function ({ userId }) {
  * @property {function} getByIds
  * @property {function} getBySamlId
  * @property {function} getByUsernameOrEmailWithRolesAndPassword
- * @property {function} getUserDetailsForAdmin
  * @property {function} isUserAllowedToAccessCertificationCenter
  * @property {function} getWithMemberships
  * @property {function} isUserExistingByEmail
@@ -438,7 +367,6 @@ export {
   getByIds,
   getBySamlId,
   getByUsernameOrEmailWithRolesAndPassword,
-  getUserDetailsForAdmin,
   getWithMemberships,
   isUserAllowedToAccessCertificationCenter,
   isUserExistingByEmail,
@@ -455,87 +383,6 @@ export {
   updateUsername,
   updateWithEmailConfirmed,
 };
-
-function _fromKnexDTOToUserDetailsForAdmin({
-  userDTO,
-  organizationLearnersDTO,
-  authenticationMethodsDTO,
-  pixAdminRolesDTO,
-  lastUserApplicationConnectionsDTO,
-}) {
-  const organizationLearners = organizationLearnersDTO.map(
-    (organizationLearnerDTO) =>
-      new OrganizationLearnerForAdmin({
-        id: organizationLearnerDTO.id,
-        firstName: organizationLearnerDTO.firstName,
-        lastName: organizationLearnerDTO.lastName,
-        birthdate: organizationLearnerDTO.birthdate,
-        division: organizationLearnerDTO.division,
-        group: organizationLearnerDTO.group,
-        organizationId: organizationLearnerDTO.organizationId,
-        organizationName: organizationLearnerDTO.organizationName,
-        createdAt: organizationLearnerDTO.createdAt,
-        updatedAt: organizationLearnerDTO.updatedAt,
-        isDisabled: organizationLearnerDTO.isDisabled,
-        organizationIsManagingStudents:
-          organizationLearnerDTO.organizationIsManagingStudents || organizationLearnerDTO.hasImportFeature,
-        additionalInformations: organizationLearnerDTO.attributes,
-        additionalColumns: organizationLearnerDTO.additionalColumns,
-      }),
-  );
-  const userLogin = new UserLogin({
-    id: userDTO.userLoginId,
-    userId: userDTO.userId,
-    failureCount: userDTO.failureCount,
-    temporaryBlockedUntil: userDTO.temporaryBlockedUntil,
-    blockedAt: userDTO.blockedAt,
-  });
-
-  const lastApplicationConnections = lastUserApplicationConnectionsDTO.map(
-    (lastUserApplicationConnectionDTO) => new LastUserApplicationConnection(lastUserApplicationConnectionDTO),
-  );
-
-  const authenticationMethods = authenticationMethodsDTO.map((authenticationMethod) => {
-    const isPixAuthenticationMethodWithAuthenticationComplement =
-      authenticationMethod.identityProvider === NON_OIDC_IDENTITY_PROVIDERS.PIX.code &&
-      authenticationMethod.authenticationComplement;
-    if (isPixAuthenticationMethodWithAuthenticationComplement) {
-      // eslint-disable-next-line no-unused-vars -- extract password so that it's not returned/displayed
-      const { password, ...authenticationComplement } = authenticationMethod.authenticationComplement;
-      return {
-        ...authenticationMethod,
-        authenticationComplement,
-      };
-    }
-
-    return authenticationMethod;
-  });
-
-  return new UserDetailsForAdmin({
-    id: userDTO.id,
-    firstName: userDTO.firstName,
-    lastName: userDTO.lastName,
-    username: userDTO.username,
-    email: userDTO.email,
-    pixCertifTermsOfServiceAccepted: userDTO.pixCertifTermsOfServiceAccepted,
-    lang: userDTO.lang,
-    locale: userDTO.locale,
-    lastPixCertifTermsOfServiceValidatedAt: userDTO.lastPixCertifTermsOfServiceValidatedAt,
-    lastLoggedAt: userDTO.lastLoggedAt,
-    emailConfirmedAt: userDTO.emailConfirmedAt,
-    organizationLearners,
-    authenticationMethods,
-    userLogin,
-    hasBeenAnonymised: userDTO.hasBeenAnonymised,
-    hasBeenAnonymisedBy: userDTO.hasBeenAnonymisedBy,
-    updatedAt: userDTO.updatedAt,
-    createdAt: userDTO.createdAt,
-    anonymisedByFirstName: userDTO.anonymisedByFirstName,
-    anonymisedByLastName: userDTO.anonymisedByLastName,
-    isPixAgent: pixAdminRolesDTO.length > 0,
-    lastApplicationConnections,
-  });
-}
 
 /**
  * @param userDTO
