+++ api/tests/deprecated/integration/domain/usecases/get-user-details-for-admin.usecase.test.js
@@ -1,10 +1,10 @@
-import { UserDetailsForAdmin } from '../../../../../src/identity-access-management/domain/models/UserDetailsForAdmin.js';
-import { usecases } from '../../../../../src/identity-access-management/domain/usecases/index.js';
+import { UserDetailsForAdmin } from '../../../../../src/deprecated/domain/models/UserDetailsForAdmin.js';
+import { usecases } from '../../../../../src/deprecated/domain/usecases/index.js';
 import { UserNotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder } from '../../../../tooling/databases.js';
 
-describe('Integration | Identity Access Management | Domain | UseCase | get-user-details-for-admin', function () {
+describe('Integration | Deprecated | Domain | UseCase | get-user-details-for-admin', function () {
   describe('#getUserDetailsForAdmin', function () {
     it('returns the found user', async function () {
       // given
