+++ api/tests/deprecated/unit/infrastructure/serializers/jsonapi/user-details-for-admin.serializer.test.js
@@ -1,10 +1,10 @@
+import { userDetailsForAdminSerializer } from '../../../../../../src/deprecated/infrastructure/serializers/jsonapi/user-details-for-admin.serializer.js';
 import { LastUserApplicationConnection } from '../../../../../../src/identity-access-management/domain/models/LastUserApplicationConnection.js';
-import { userDetailsForAdminSerializer } from '../../../../../../src/identity-access-management/infrastructure/serializers/jsonapi/user-details-for-admin.serializer.js';
 import { STATUS } from '../../../../../../src/legal-documents/domain/models/LegalDocumentStatus.js';
 import { expect } from '../../../../../test-helper.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
 
-describe('Unit | Serializer | JSONAPI | user-details-for-admin-serializer', function () {
+describe('Unit | Deprecated | Infrastructure | Serializer | JSONAPI | user-details-for-admin-serializer', function () {
   describe('#serialize', function () {
     it('serializes user details for Pix Admin', function () {
       // given
